3.1-drf-intro/smart_home/requirements.txt
django 3.7.2
djangorestframework 3.14.0
psycopg2-binary 2.9.5
